# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Admin System API",
    "version": "1.0.0",
    "description": "后台管理系统 API 文档 - 自动生成"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
...
  "paths": {
...
    "/manage/sys/role/getRoleAdminList": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "获取角色列表（分页）",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "page": {
                    "type": "number",
                    "description": "页码",
                    "example": 1
                  },
                  "rows": {
                    "type": "number",
                    "description": "每页数量",
                    "example": 10
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/role/getAllRoles": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "获取所有角色（不分页）",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/manage/sys/role/addRole": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "添加角色",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roleName",
                  "roleCode"
                ],
                "properties": {
                  "roleName": {
                    "type": "string",
                    "description": "角色名称",
                    "example": "管理员"
                  },
                  "roleCode": {
                    "type": "string",
                    "description": "角色编码",
                    "example": "admin"
                  },
                  "description": {
                    "type": "string",
                    "description": "描述",
                    "example": "系统管理员"
                  },
                  "sortOrder": {
                    "type": "number",
                    "description": "排序",
                    "example": 1
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/role/editRole": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "编辑角色",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "角色ID",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "roleName": {
                    "type": "string",
                    "description": "角色名称",
                    "example": "管理员"
                  },
                  "roleCode": {
                    "type": "string",
                    "description": "角色编码",
                    "example": "admin"
                  },
                  "description": {
                    "type": "string",
                    "description": "描述",
                    "example": "系统管理员"
                  },
                  "sortOrder": {
                    "type": "number",
                    "description": "排序",
                    "example": 1
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/role/deleteRole": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "删除角色",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "角色ID",
                    "example": "507f1f77bcf86cd799439011"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/role/getRoleUserList": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "获取角色用户列表",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roleId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/role/UpdateRoleUserList": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "更新角色用户",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roleId": {
                    "example": "any"
                  },
                  "userIds": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}